+++ src/Structure/Field/LinkField.php
@@ -2,8 +2,11 @@
 
 namespace Torr\PrismicApi\Structure\Field;
 
+use Torr\PrismicApi\Data\Value\DocumentLinkValue;
+use Torr\PrismicApi\Data\Value\ImageValue;
 use Torr\PrismicApi\Exception\Structure\InvalidTypeDefinitionException;
 use Torr\PrismicApi\Structure\Helper\FilterFieldsHelper;
+use Torr\PrismicApi\Transform\FieldValueTransformer;
 
 /**
  * @see https://prismic.io/docs/core-concepts/link-content-relationship
@@ -29,14 +32,14 @@ final class LinkField extends InputField
 	 */
 	public function __construct (
 		string $label,
-		?string $select = self::SELECT_ALL,
+		private ?string $select = self::SELECT_ALL,
 		?string $placeholder = null,
 		?array $customTypes = null,
 		?array $tags = null,
 	)
 	{
 		$hasDocumentFilter = !empty($customTypes) || !empty($tags);
-		$selectsDocuments = self::SELECT_ALL === $select || self::SELECT_DOCUMENT === $select;
+		$selectsDocuments = self::SELECT_ALL === $this->select || self::SELECT_DOCUMENT === $this->select;
 
 		if ($hasDocumentFilter && !$selectsDocuments)
 		{
@@ -47,7 +50,7 @@ final class LinkField extends InputField
 		parent::__construct(self::TYPE_KEY, FilterFieldsHelper::filterOptionalFields([
 			"label" => $label,
 			"placeholder" => $placeholder,
-			"select" => $select,
+			"select" => $this->select,
 			"customtypes" => $customTypes,
 			"tags" => $tags,
 		]));
@@ -60,5 +63,46 @@ final class LinkField extends InputField
 	{
 		// @todo add validation
 		return [];
+	}
+
+
+	/**
+	 * @inheritDoc
+	 */
+	public function transformValue (mixed $data, FieldValueTransformer $valueTransformer) : mixed
+	{
+		$type = $data["link_type"] ?? null;
+
+		if ("Web" === $type)
+		{
+			return $data["url"] ?? null;
+		}
+
+		if ("Media" === $type)
+		{
+			// return as an image, if specifically an image was asked for
+			// @todo always return it this way (and use ImageValue or FileValue)
+			if (self::SELECT_MEDIA === $this->select)
+			{
+				return new ImageValue(
+					$data["url"],
+					(int) $data["width"],
+					(int) $data["height"],
+					$data["name"],
+				);
+			}
+
+			return $data["url"] ?? null;
+		}
+
+		if ("Document" === $type && \is_string($data["id"]))
+		{
+			return new DocumentLinkValue(
+				$data["id"],
+				$data["type"] ?? null,
+			);
+		}
+
+		return null;
 	}
 }